core: Properly track whether we should actually remove the PID file
Previously, PyLink spuriously removed PID files even if -n/--no-check-pid was set or if PID file checking caused PyLink to quit

(cherry picked from commit 91659ea992e5be9fc645f8805a09c2dffef4148e)

+++ coremods/control.py
@@ -24,13 +24,16 @@ def _print_remaining_threads():
     log.debug('_shutdown(): Remaining threads: %s', ['%s/%s' % (t.name, t.ident) for t in threading.enumerate()])
 
 def _remove_pid():
-    # Remove our pid file.
     pidfile = "%s.pid" % conf.confname
-    log.info("Removing PID file %r.", pidfile)
-    try:
-        os.remove(pidfile)
-    except OSError:
-        log.exception("Failed to remove PID file %r, ignoring..." % pidfile)
+    if world._should_remove_pid:
+        # Remove our pid file.
+        log.info("Removing PID file %r.", pidfile)
+        try:
+            os.remove(pidfile)
+        except OSError:
+            log.exception("Failed to remove PID file %r, ignoring..." % pidfile)
+    else:
+        log.debug('Not removing PID file %s as world._should_remove_pid is False.' % pidfile)
 
 def _kill_plugins(irc=None):
     log.info("Shutting down plugins.")
